+++ src/commands/build.rs
@@ -1,4 +1,5 @@
 use crate::libs::helpers::{clear_console, spinner_start, convert_bytes};
+use crate::libs::project_config::ProjectConfig;
 use crate::tools::webpack;
 use clap::Args;
 use serde_json::{from_str, Value};
@@ -37,7 +38,8 @@ pub fn cmd(build_args: BuildArgs) -> Result<(), Box<dyn Error>> {
 
     let mut child_stdin: std::process::ChildStdin = child.stdin.take().expect("Failed to open stdin for child process");
     let mut spinner = spinner_start("Loading...").unwrap();
-    let webpack_config = webpack::get_config_prod(true);
+    let project_config = ProjectConfig::get(&None);
+    let webpack_config = webpack::get_config_prod(true, &project_config);
     let json_string = serde_json::to_string(&webpack_config).unwrap();
     child_stdin.write_all(&json_string.as_bytes()).expect("Failed to write to child process stdin");
     drop(child_stdin);