+++ src/libs/project_config.rs
@@ -1,7 +1,10 @@
 use super::helpers::is_default;
+use crate::commands::start::StartArgs;
+use crate::templates::{Framework, ProjectProps};
 use serde::{Deserialize, Serialize};
+use serde_json::{from_str, to_string};
 use std::error::Error;
-use std::fs::File;
+use std::fs::{read_to_string, File};
 
 pub const PROJECT_CONFIG: &str = "lyrn.json";
 
@@ -34,6 +37,46 @@ impl ProjectConfig {
         serde_json::to_writer_pretty(&file, &self)?;
         Ok(())
     }
+
+    pub fn get(start_args: &Option<StartArgs>) -> Self {
+        let data = read_to_string(PROJECT_CONFIG).unwrap_or(to_string(&Self::default()).unwrap());
+        let mut project_config: Self = from_str(&data).unwrap_or_default();
+        if start_args.is_some() {
+            let start_args = start_args.clone().unwrap();
+            if start_args.port.is_some() {
+                project_config.dev.port = start_args.port.unwrap()
+            }
+        }
+        project_config
+    }
+
+    pub fn create(project_props: &ProjectProps) -> Self {
+        let mut project_config = Self::default();
+        project_config.app = AppConfig {
+            name: project_props.clone().name,
+            title: project_props.name.to_uppercase(),
+            framework: project_props.framework,
+        };
+        project_config
+    }
+
+    pub fn default() -> Self {
+        Self {
+            app: AppConfig { ..Default::default() },
+            dev: DevConfig {
+                public_path: "/".into(),
+                protocol: "http".into(),
+                host: "localhost".into(),
+                port: 8080,
+                ..Default::default()
+            },
+            prod: ProdConfig {
+                public_path: "/".into(),
+                ..Default::default()
+            },
+            ..Default::default()
+        }
+    }
 }
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
@@ -42,6 +85,8 @@ pub struct AppConfig {
     pub name: String,
     #[serde(default, skip_serializing_if = "is_default")]
     pub title: String,
+    #[serde(default, skip_serializing_if = "is_default")]
+    pub framework: Framework,
 }
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
@@ -64,26 +109,4 @@ pub struct ProdConfig {
     pub public_path: String,
     #[serde(default, skip_serializing_if = "is_default")]
     pub config: String,
-}
-
-pub fn project_config() -> ProjectConfig {
-    ProjectConfig {
-        app: AppConfig {
-            name: "app".into(),
-            title: "New application".into(),
-            ..Default::default()
-        },
-        dev: DevConfig {
-            public_path: "/".into(),
-            protocol: "http".into(),
-            host: "localhost".into(),
-            port: 8080,
-            ..Default::default()
-        },
-        prod: ProdConfig {
-            public_path: "/".into(),
-            ..Default::default()
-        },
-        ..Default::default()
-    }
 }